chore: Support widgetization in breadcrumb-group, drawer, help-panel, side-navigation components (#2234)
