+++ src/breadcrumb-group/utils.ts
@@ -1,0 +1,9 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+import { BreadcrumbGroupProps } from './interfaces';
+
+export const getEventDetail = <T extends BreadcrumbGroupProps.Item>(item: T) => ({
+  item,
+  text: item.text,
+  href: item.href,
+});
