+++ src/side-navigation/index.tsx
@@ -1,78 +1,16 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import React, { useMemo, useCallback, useEffect } from 'react';
-import clsx from 'clsx';
-import { getBaseProps } from '../internal/base-component';
-import { fireNonCancelableEvent, fireCancelableEvent } from '../internal/events';
+import React from 'react';
 import { SideNavigationProps } from './interfaces';
-import { Header, NavigationItemsList } from './internal';
-import { generateExpandableItemsMapping, checkDuplicateHrefs } from './util';
-import styles from './styles.css.js';
-import { isDevelopment } from '../internal/is-development';
 import { applyDisplayName } from '../internal/utils/apply-display-name';
 import useBaseComponent from '../internal/hooks/use-base-component';
+import { InternalSideNavigation } from './internal';
 
 export { SideNavigationProps };
 
-export default function SideNavigation({
-  header,
-  activeHref,
-  items = [],
-  onFollow,
-  onChange,
-  ...props
-}: SideNavigationProps) {
-  const { __internalRootRef } = useBaseComponent('SideNavigation');
-  const baseProps = getBaseProps(props);
-  const parentMap = useMemo(() => generateExpandableItemsMapping(items), [items]);
-
-  if (isDevelopment) {
-    // This code should be wiped in production anyway.
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    useEffect(() => checkDuplicateHrefs(items), [items]);
-  }
-
-  const onChangeHandler = useCallback(
-    (item: SideNavigationProps.Section | SideNavigationProps.ExpandableLinkGroup, expanded: boolean) => {
-      // generateExpandableItemsMapping walks through the entire tree, so we're certain about getting a value.
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      fireNonCancelableEvent(onChange, { item, expanded: expanded, expandableParents: parentMap.get(item)! });
-    },
-    [onChange, parentMap]
-  );
-
-  const onFollowHandler = useCallback(
-    (
-      item:
-        | SideNavigationProps.Link
-        | SideNavigationProps.Header
-        | SideNavigationProps.LinkGroup
-        | SideNavigationProps.ExpandableLinkGroup,
-      sourceEvent: React.SyntheticEvent | Event
-    ) => {
-      fireCancelableEvent(onFollow, item, sourceEvent);
-    },
-    [onFollow]
-  );
-
-  return (
-    <div {...baseProps} className={clsx(styles.root, baseProps.className)} ref={__internalRootRef}>
-      {header && (
-        <Header definition={header} activeHref={activeHref} fireChange={onChangeHandler} fireFollow={onFollowHandler} />
-      )}
-      {items && (
-        <div className={styles['list-container']}>
-          <NavigationItemsList
-            variant="root"
-            items={items}
-            fireFollow={onFollowHandler}
-            fireChange={onChangeHandler}
-            activeHref={activeHref}
-          />
-        </div>
-      )}
-    </div>
-  );
+export default function SideNavigation({ items = [], ...props }: SideNavigationProps) {
+  const internalProps = useBaseComponent('SideNavigation');
+  return <InternalSideNavigation {...props} {...internalProps} items={items} />;
 }
 
 applyDisplayName(SideNavigation, 'SideNavigation');
